Move and re-namespace HeadVersion
